add Question05

+++ src/Questao04.java
@@ -5,11 +5,14 @@ public class Questao04 {
         Scanner ler = new Scanner(System.in);
         System.out.println("Digite um Nome do trabalhador: ");
         String name = ler.next();
+
         System.out.println("Digite o valor da hora do trabalhador " + name + " R$ ");
-        float vlrHora = ler.nextFloat();
+        double vlrHora = ler.nextDouble();
+
         System.out.println("Digite quantas  horas o " + name + " trabalhou");
-        float qtdHora = ler.nextFloat();
-        float salarioFinal = (vlrHora * qtdHora);
+        double qtdHora = ler.nextDouble();
+        
+        double salarioFinal = (vlrHora * qtdHora);
         
         System.out.println("O Salário do " + name + " é  R$ " + salarioFinal);
